Add support for negative chunk nums

+++ MinecraftClone/stdafx.h
@@ -13,7 +13,7 @@ constexpr size_t TEX_SIZE = 8;
 
 constexpr size_t NUM_BLOCKS = 3;
 
-constexpr size_t CHUNK_X_LEN = 16;
-constexpr size_t CHUNK_Y_LEN = 16;
-constexpr size_t CHUNK_Z_LEN = 16;
-constexpr size_t CHUNK_VOLUME = CHUNK_X_LEN * CHUNK_Y_LEN * CHUNK_Z_LEN;
+constexpr int CHUNK_X_LEN = 16;
+constexpr int CHUNK_Y_LEN = 16;
+constexpr int CHUNK_Z_LEN = 16;
+constexpr int CHUNK_VOLUME = CHUNK_X_LEN * CHUNK_Y_LEN * CHUNK_Z_LEN;
